Change domain data type

+++ MailChimp.Api.Net/Domain/Reports/ListStats.cs
@@ -8,8 +8,8 @@ namespace MailChimp.Api.Net.Domain.Reports
 {
     public class ListStats
     {
-        public int sub_rate { get; set; }
-        public int unsub_rate { get; set; }
+        public double sub_rate { get; set; }
+        public double unsub_rate { get; set; }
         public double open_rate { get; set; }
         public double click_rate { get; set; }
     }
